+++ tests/Classifiers/LogisticRegressionTest.php
@@ -9,6 +9,7 @@ use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Probabilistic;
 use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Other\Loggers\BlackHole;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
 use Rubix\ML\Classifiers\LogisticRegression;
@@ -37,6 +38,8 @@ class LogisticRegressionTest extends TestCase
         ], [0.45, 0.55]);
 
         $this->estimator = new LogisticRegression(10, new Adam(0.01), 1e-4, 300, 1e-4, new CrossEntropy());
+
+        $this->estimator->setLogger(new BlackHole());
     }
 
     public function test_build_classifier()