+++ tests/Other/Strategies/MeanTest.php
@@ -2,12 +2,12 @@
 
 namespace Rubix\ML\Tests\Other\Strategies;
 
+use Rubix\ML\Other\Strategies\Mean;
 use Rubix\ML\Other\Strategies\Strategy;
 use Rubix\ML\Other\Strategies\Continuous;
-use Rubix\ML\Other\Strategies\BlurryMean;
 use PHPUnit\Framework\TestCase;
 
-class BlurryMeanTest extends TestCase
+class MeanTest extends TestCase
 {
     protected $values;
 
@@ -21,12 +21,12 @@ class BlurryMeanTest extends TestCase
 
         $this->range = [0, 6];
 
-        $this->strategy = new BlurryMean(0.3);
+        $this->strategy = new Mean();
     }
 
     public function test_build_strategy()
     {
-        $this->assertInstanceOf(BlurryMean::class, $this->strategy);
+        $this->assertInstanceOf(Mean::class, $this->strategy);
         $this->assertInstanceOf(Continuous::class, $this->strategy);
         $this->assertInstanceOf(Strategy::class, $this->strategy);
     }
@@ -37,9 +37,6 @@ class BlurryMeanTest extends TestCase
 
         $guess = $this->strategy->guess();
 
-        $this->assertThat($guess, $this->logicalAnd(
-            $this->greaterThanOrEqual($this->range[0]),
-            $this->lessThanOrEqual($this->range[1]))
-        );
+        $this->assertEquals(3., $guess);
     }
 }