+++ CHANGELOG.md
@@ -17,6 +17,8 @@
     - Classification and Regression trees now Verbose
     - Random Forest can now return feature importances
     - Gradient Boost now accepts base and booster estimators
+    - Blurry Median strategy is now Blurry Percentile
+    - Added Mean strategy
 
 - 0.0.6-alpha
     - Added Gradient Boost regressor